pipeline/requirements.txt
# ign-data-platform/pipeline/requirements.txt
# Python dependencies for the Airflow pipeline
# Installed inside the custom Docker image at build time

# AWS SDK — needed for S3 upload and Glue crawler trigger
boto3>=1.34.0
botocore>=1.34.0

# HTTP requests — YouTube API calls
requests>=2.31.0

# Airflow providers
apache-airflow-providers-amazon>=8.0.0
